feat(wallet): change code style
