+++ src/components/CoboVault/CoboVaultExtendedPublicKeyImporter.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import PropTypes from "prop-types";
 import { CoboVaultExportExtendedPublicKey } from "unchained-wallets";
 
@@ -7,80 +7,68 @@ import { FormGroup, FormHelperText } from "@material-ui/core";
 
 import { CoboVaultReader } from "./index";
 
-class CoboVaultExtendedPublicKeyImporter extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      extendedPublicKeyError: "",
-    };
-  }
-
-  componentDidMount = () => {
-    const { resetBIP32Path } = this.props;
+const CoboVaultExtendedPublicKeyImporter = (props) => {
+  const {
+    resetBIP32Path,
+    network,
+    extendedPublicKeyImporter,
+    disableChangeMethod,
+    validateAndSetBIP32Path,
+    validateAndSetExtendedPublicKey,
+    enableChangeMethod,
+    validateAndSetRootFingerprint,
+    reset,
+  } = props;
+  const [extendedPublicKeyError, setExtendedPublicKeyError] = useState("");
+  useEffect(() => {
     resetBIP32Path();
-  };
+  }, [resetBIP32Path]);
 
-  interaction = () => {
-    const { network, extendedPublicKeyImporter } = this.props;
+  const interaction = () => {
     return new CoboVaultExportExtendedPublicKey({
       network,
       bip32Path: extendedPublicKeyImporter.bip32Path,
     });
   };
 
-  render = () => {
-    const { disableChangeMethod } = this.props;
-    const { extendedPublicKeyError } = this.state;
-    return (
-      <FormGroup>
-        <CoboVaultReader
-          qrStartText="Import Extended Public Key"
-          fileStartText="Upload Extended Public Key"
-          interaction={this.interaction()}
-          onStart={disableChangeMethod}
-          onSuccess={this.import}
-          onClear={this.onClear}
-          shouldShowFileReader
-          fileType="json"
-        />
-        <FormHelperText className="text-danger">
-          {extendedPublicKeyError}
-        </FormHelperText>
-      </FormGroup>
-    );
-  };
-
-  setError = (value) => {
-    this.setState({ extendedPublicKeyError: value });
-  };
-
-  import = (data) => {
-    const {
-      validateAndSetBIP32Path,
-      validateAndSetExtendedPublicKey,
-      enableChangeMethod,
-      validateAndSetRootFingerprint,
-    } = this.props;
+  const importExtendedPublicKey = (data) => {
     enableChangeMethod();
     const { xpub, bip32Path, rootFingerprint: fingerprint } = data;
     validateAndSetBIP32Path(
       bip32Path,
       () => {
-        validateAndSetExtendedPublicKey(xpub, this.setError, () => {
+        validateAndSetExtendedPublicKey(xpub, setExtendedPublicKeyError, () => {
           validateAndSetRootFingerprint(fingerprint);
         });
       },
-      this.setError
+      setExtendedPublicKeyError
     );
   };
 
-  onClear = () => {
-    const { reset, enableChangeMethod } = this.props;
+  const onClear = () => {
     reset(true); // clear BIP32 path
-    this.setError("");
+    setExtendedPublicKeyError("");
     enableChangeMethod();
   };
-}
+
+  return (
+    <FormGroup>
+      <CoboVaultReader
+        qrStartText="Import Extended Public Key"
+        fileStartText="Upload Extended Public Key"
+        interaction={interaction()}
+        onStart={disableChangeMethod}
+        onSuccess={importExtendedPublicKey}
+        onClear={onClear}
+        shouldShowFileReader
+        fileType="json"
+      />
+      <FormHelperText className="text-danger">
+        {extendedPublicKeyError}
+      </FormHelperText>
+    </FormGroup>
+  );
+};
 
 CoboVaultExtendedPublicKeyImporter.propTypes = {
   enableChangeMethod: PropTypes.func.isRequired,
