+++ src/components/CoboVault/CoboVaultReader.jsx
@@ -1,114 +1,93 @@
-import React, { Component } from "react";
+import React, { useState } from "react";
 import PropTypes from "prop-types";
-import { PENDING, ACTIVE } from "unchained-wallets";
-import "./styles.css";
+import { PENDING } from "unchained-wallets";
 
 import { Button, FormHelperText, Grid } from "@material-ui/core";
 
 import CoboVaultRawReader from "./CoboVaultRawReader";
 
-class CoboVaultReader extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      status: PENDING,
-      error: "",
-    };
+const CoboVaultReader = (props) => {
+  const [status, setStatus] = useState(PENDING);
+  const [error, setError] = useState("");
+  const {
+    interaction,
+    qrStartText,
+    fileStartText,
+    shouldShowFileReader,
+    fileType,
+    onClear,
+    onSuccess,
+  } = props;
+
+  const handleError = (err) => {
+    setStatus("error");
+    setError(err.message);
+    if (onClear) {
+      onClear();
+    }
+  };
+
+  const handleSuccess = (result) => {
+    onSuccess(result);
+    setStatus("success");
+  };
+
+  const handleStop = () => {
+    setStatus(PENDING);
+    setError("");
+    if (onClear) {
+      onClear();
+    }
+  };
+
+  if (status === PENDING) {
+    const commandMessage = interaction.messageFor({
+      state: status,
+      code: "cobo.command",
+    });
+    return (
+      <div>
+        <p>{commandMessage.instructions}</p>
+        <p>
+          When you are ready, scan the QR code produced by Cobo Vault or upload
+          the file generated by Cobo Vault:
+        </p>
+        <Grid container spacing={2}>
+          <CoboVaultRawReader
+            shouldShowFileReader={shouldShowFileReader}
+            fileStartText={fileStartText}
+            qrStartText={qrStartText}
+            interaction={interaction}
+            handleError={handleError}
+            handleSuccess={handleSuccess}
+            handleStop={handleStop}
+            fileType={fileType}
+          />
+        </Grid>
+      </div>
+    );
   }
 
-  render = () => {
-    const { status, error } = this.state;
-    const {
-      interaction,
-      qrStartText,
-      fileStartText,
-      shouldShowFileReader,
-      fileType,
-    } = this.props;
+  if (status === "error" || status === "success") {
+    return (
+      <div>
+        <FormHelperText error>{error}</FormHelperText>
+        <Button
+          variant="contained"
+          color="secondary"
+          size="small"
+          onClick={handleStop}
+        >
+          Reset
+        </Button>
+      </div>
+    );
+  }
 
-    if (status === PENDING) {
-      const commandMessage = interaction.messageFor({
-        state: status,
-        code: "cobo.command",
-      });
-      return (
-        <div>
-          <p>{commandMessage.instructions}</p>
-          <p>
-            When you are ready, scan the QR code produced by Cobo Vault or
-            upload the file generated by Cobo Vault:
-          </p>
-          <Grid container spacing={2}>
-            <CoboVaultRawReader
-              shouldShowFileReader={shouldShowFileReader}
-              fileStartText={fileStartText}
-              qrStartText={qrStartText}
-              interaction={interaction}
-              handleError={this.handleError}
-              handleSuccess={this.handleSuccess}
-              handleStop={this.handleStop}
-              fileType={fileType}
-            />
-          </Grid>
-        </div>
-      );
-    }
-
-    if (status === "error" || status === "success") {
-      return (
-        <div>
-          <FormHelperText error>{error}</FormHelperText>
-          <Button
-            variant="contained"
-            color="secondary"
-            size="small"
-            onClick={this.handleStop}
-          >
-            Reset
-          </Button>
-        </div>
-      );
-    }
-
-    return null;
-  };
-
-  handleStart = () => {
-    const { onStart } = this.props;
-    this.setState({ status: ACTIVE, error: "" });
-    if (onStart) {
-      onStart();
-    }
-  };
-
-  handleError = (error) => {
-    const { onClear } = this.props;
-    this.setState({ status: "error", error: error.message });
-    if (onClear) {
-      onClear();
-    }
-  };
-
-  handleSuccess = (result) => {
-    const { onSuccess } = this.props;
-    onSuccess(result);
-    this.setState({ status: "success" });
-  };
-
-  handleStop = () => {
-    const { onClear } = this.props;
-    this.setState({
-      status: PENDING,
-      error: "",
-    });
-    if (onClear) {
-      onClear();
-    }
-  };
-}
+  return null;
+};
 
 CoboVaultReader.propTypes = {
-  onStart: PropTypes.func.isRequired,
   onSuccess: PropTypes.func.isRequired,
   onClear: PropTypes.func.isRequired,
   qrStartText: PropTypes.string,
