--- src/components/CoboVault/styles.css
@@ -1,42 +1,0 @@
-.qr-player__container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 16px;
-  background: #FAFAFA;
-  width: 480px;
-  margin-left: auto;
-  margin-right: auto;
-  margin-top: 50px;
-  outline: none;
-}
-
-.qr-player__title {
-  font-size: 16px;
-  font-weight: 500;
-}
-
-.qr-player__description {
-
-}
-
-.qr-player__qr-container {
-  display: flex;
-  width: 360px;
-  height: 360px;
-  align-items: center;
-  justify-content: center;
-}
-
-.qr-scanner__container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 16px;
-  background: #FAFAFA;
-  width: 600px;
-  margin-left: auto;
-  margin-right: auto;
-  margin-top: 50px;
-  outline: none;
-}
