+++ src/components/QR/QR.module.scss
@@ -1,0 +1,42 @@
+.qr-player {
+  &__container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 16px;
+    background: #fafafa;
+    width: 480px;
+    margin-left: auto;
+    margin-right: auto;
+    margin-top: 50px;
+    outline: none;
+  }
+
+  &__title {
+    font-size: 16px;
+    font-weight: 500;
+  }
+
+  &__qr-container {
+    display: flex;
+    width: 360px;
+    height: 360px;
+    align-items: center;
+    justify-content: center;
+  }
+}
+
+.qr-scanner {
+  &__container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 16px;
+    background: #fafafa;
+    width: 600px;
+    margin-left: auto;
+    margin-right: auto;
+    margin-top: 50px;
+    outline: none;
+  }
+}
